+++ spec/shared_helper.rb
@@ -5,6 +5,8 @@ require 'gamefic'
 
 require_relative 'fixtures/plot_with_chapter'
 
+Gamefic.logger.level = Logger::WARN
+
 RSpec.configure do |config|
   # Run specs in random order to surface order dependencies. If you find an
   # order dependency and want to debug it, you can fix the order by providing