+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Normalized arguments accept strings
 - Smarter picks and proxies
 - Commands prefer strictness over precision
+- Queries scan for ambiguity before filtering through arguments
 
 ## 3.5.0 - October 5, 2024
 - Configurable scanners